refactor: Scheduler를 추상화하여 Presenter에 주입
