+++ app/src/main/java/woowacourse/movie/domain/scheduler/Scheduler.kt
@@ -1,0 +1,18 @@
+package woowacourse.movie.domain.scheduler
+
+import java.time.LocalDate
+import java.time.LocalDateTime
+import java.time.LocalTime
+
+interface Scheduler {
+    fun getScreeningDates(
+        startDate: LocalDate,
+        endDate: LocalDate,
+        today: LocalDate,
+    ): List<LocalDate>
+
+    fun getShowtimes(
+        selectedDate: LocalDate,
+        now: LocalDateTime,
+    ): List<LocalTime>
+}
